+++ src/Webiny/Component/Crypt/Tests/CryptTest.php
@@ -50,7 +50,7 @@ class CryptTest extends \PHPUnit_Framework_TestCase
     public function testGenerateRandomString()
     {
         $crypt = new Crypt();
-        $randomString = $crypt->generateRandomString(9, $chars = 'abc');
+        $randomString = $crypt->generateRandomString(9, 'abc');
 
         $this->assertInternalType('string', $randomString);
     }
@@ -64,14 +64,6 @@ class CryptTest extends \PHPUnit_Framework_TestCase
         $this->assertSame(64, $size);
     }
 
-    public function testGenerateUserReadableString2()
-    {
-        $crypt = new Crypt();
-        $randomString = $crypt->generateUserReadableString('asd');
-
-        $size = strlen($randomString);
-        $this->assertSame(0, $size);
-    }
 
     public function testGenerateHardReadableString()
     {
@@ -108,34 +100,27 @@ class CryptTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException \Webiny\Component\Crypt\CryptException
-     * @expectedExceptionMessage The supplied key block is in the valid sizes
-     */
-    public function testEncrypt()
-    {
-        $crypt = new Crypt();
-        $crypt->encrypt('some string', 'too short key');
-    }
-
-    /**
-     * @expectedException \Webiny\Component\Crypt\CryptException
-     * @expectedExceptionMessage Supplied Initialization Vector is too short
-     */
-    public function testEncrypt2()
-    {
-        $crypt = new Crypt();
-        $crypt->encrypt('some string', 'abcdefgh12345678', 'too short');
-    }
-
-    /**
      * @dataProvider encryptDescryptDataProvider
      */
-    public function testEncryptDecrypt($stringToEncrypt, $encKey, $encInitValue, $decKey, $decInitValue, $result)
+    public function testEncryptDecrypt($stringToEncrypt, $encKey, $decKey, $result)
     {
         $crypt = new Crypt();
-        $encrypted = $crypt->encrypt($stringToEncrypt, $encKey, $encInitValue);
+        $encrypted = $crypt->encrypt($stringToEncrypt, $encKey);
 
-        $this->assertSame($result, $crypt->decrypt($encrypted, $decKey, $decInitValue));
+        $this->assertSame($result, $crypt->decrypt($encrypted, $decKey));
+    }
+
+    /**
+     * @throws \Webiny\Component\Crypt\CryptException
+     * @expectedExceptionMessage Unable to decrypt the data.
+     * @expectedException \Webiny\Component\Crypt\CryptException
+     */
+    public function testEncryptDecryptFail()
+    {
+        $crypt = new Crypt();
+        $encrypted = $crypt->encrypt('test string', 'some key');
+
+        $crypt->decrypt($encrypted, 'wrong key');
     }
 
     public function encryptDescryptDataProvider()
@@ -145,61 +130,33 @@ class CryptTest extends \PHPUnit_Framework_TestCase
             [
                 'a',
                 'abcdefgh12345678',
-                'init_vector',
                 'abcdefgh12345678',
-                'init_vector',
                 'a'
             ],
             [
                 'B',
                 '12345678abcdefgh',
-                'init_vector',
                 '12345678abcdefgh',
-                'init_vector',
                 'B'
             ],
             [
                 'test string',
                 'abcdefgh12345678',
-                'init_vector',
                 'abcdefgh12345678',
-                'init_vector',
                 'test string'
             ],
             [
                 'test string',
                 'abcdefgh/&%$#"!?',
-                'init_vector',
                 'abcdefgh/&%$#"!?',
-                'init_vector',
                 'test string'
             ],
             [
                 'test string',
                 '-.,_:;></&%$#"!?',
-                'init_vector',
                 '-.,_:;></&%$#"!?',
-                'init_vector',
                 'test string'
-            ],
-            // decryption fails because the key doesn't match
-            [
-                'test string',
-                'abcdefgh12345678',
-                'init_vector',
-                '12345678abcdefgh',
-                'init_vector',
-                false
-            ],
-            // decryption fails because the initialization vector doesn't match
-            [
-                'test string',
-                'abcdefgh12345678',
-                'init_vector',
-                'abcdefgh12345678',
-                'foo_vector_',
-                false
-            ],
+            ]
         ];
     }
 }